Move copy/paste NotificationArea to cGame
