+++ src/game/NotificationArea.cpp
